Moved all shipping and tax functions to a new shipping and tax model, Moved getMagentoUrl from Api helper to Url helper
