+++ app/code/community/Bolt/Boltpay/Helper/Api.php
@@ -225,10 +225,10 @@ class Bolt_Boltpay_Helper_Api extends Bolt_Boltpay_Helper_Data
             ///  the Bolt transaction.
             //////////////////////////////////////////////////////////////////////////////////
             $shippingMethodCode = null;
-            $shippingAndTaxHelper = Mage::helper("boltpay/shippingAndTax");
+            $shippingAndTaxModel = Mage::getModel("boltpay/shippingAndTax");
             if ($transaction->order->cart->shipments) {
 
-                $shippingAndTaxHelper->applyShippingAddressToQuote($immutableQuote, $transaction->order->cart->shipments[0]->shipping_address);
+                $shippingAndTaxModel->applyShippingAddressToQuote($immutableQuote, $transaction->order->cart->shipments[0]->shipping_address);
                 $shippingMethodCode = $transaction->order->cart->shipments[0]->reference;
             }
 
@@ -923,150 +923,6 @@ class Bolt_Boltpay_Helper_Api extends Bolt_Boltpay_Helper_Data
     }
 
     /**
-     * Gets the shipping and the tax estimate for a quote
-     *
-     * @param Mage_Sales_Model_Quote  $quote    A quote object with pre-populated addresses
-     *
-     * @return array    Bolt shipping and tax response array to be converted to JSON
-     */
-    public function getShippingAndTaxEstimate( $quote )
-    {
-        $response = array(
-            'shipping_options' => array(),
-            'tax_result' => array(
-                "amount" => 0
-            ),
-        );
-
-        Mage::helper('boltpay')->collectTotals(Mage::getModel('sales/quote')->load($quote->getId()));
-
-        //we should first determine if the cart is virtual
-        if($quote->isVirtual()){
-            Mage::helper('boltpay')->collectTotals($quote, true);
-            $option = array(
-                "service"   => Mage::helper('boltpay')->__('No Shipping Required'),
-                "reference" => 'noshipping',
-                "cost" => 0,
-                "tax_amount" => abs(round($quote->getBillingAddress()->getTaxAmount() * 100))
-            );
-            $response['shipping_options'][] = $option;
-            $quote->setTotalsCollectedFlag(true);
-            return $response;
-        }
-
-        /*****************************************************************************************
-         * Calculate tax
-         *****************************************************************************************/
-        $this->applyShippingRate($quote, null);
-
-        $shippingAddress = $quote->getShippingAddress();
-        $shippingAddress->setCollectShippingRates(true)->collectShippingRates()->save();
-
-        $originalDiscountedSubtotal = $quote->getSubtotalWithDiscount();
-
-        $rates = $this->getSortedShippingRates($shippingAddress);
-
-        foreach ($rates as $rate) {
-
-            if ($rate->getErrorMessage()) {
-                $metaData = array('quote' => var_export($quote->debug(), true));
-                Mage::helper('boltpay/bugsnag')->notifyException(
-                    new Exception(
-                        Mage::helper('boltpay')->__("Error getting shipping option for %s: %s", $rate->getCarrierTitle(), $rate->getErrorMessage())
-                    ),
-                    $metaData
-                );
-                continue;
-            }
-
-            $this->applyShippingRate($quote, $rate->getCode());
-
-            $rateCode = $rate->getCode();
-
-            if (empty($rateCode)) {
-                $metaData = array('quote' => var_export($quote->debug(), true));
-
-                Mage::helper('boltpay/bugsnag')->notifyException(
-                    new Exception( Mage::helper('boltpay')->__('Rate code is empty. ') . var_export($rate->debug(), true) ),
-                    $metaData
-                );
-            }
-
-            $adjustedShippingAmount = $this->getAdjustedShippingAmount($originalDiscountedSubtotal, $quote);
-
-            $option = array(
-                "service" => $this->getShippingLabel($rate),
-                "reference" => $rateCode,
-                "cost" => round($adjustedShippingAmount * 100),
-                "tax_amount" => abs(round($shippingAddress->getTaxAmount() * 100))
-            );
-
-            $response['shipping_options'][] = $option;
-        }
-
-        return $response;
-    }
-
-    /**
-     * Applies shipping rate to quote. Clears previously calculated discounts by clearing address id.
-     *
-     * @param Mage_Sales_Model_Quote $quote    Quote which has been updated to use new shipping rate
-     * @param string $shippingRateCode    Shipping rate code
-     */
-    public function applyShippingRate($quote, $shippingRateCode) {
-        $shippingAddress = $quote->getShippingAddress();
-
-        if (!empty($shippingAddress)) {
-            // Flagging address as new is required to force collectTotals to recalculate discounts
-            $shippingAddress->isObjectNew(true);
-            $shippingAddressId = $shippingAddress->getData('address_id');
-
-            $shippingAddress->setShippingMethod($shippingRateCode);
-
-            // When multiple shipping methods apply a discount to the sub-total, collect totals doesn't clear the
-            // previously set discount, so the previous discount gets added to each subsequent shipping method that
-            // includes a discount. Here we reset it to the original amount to resolve this bug.
-            $quoteItems = $quote->getAllItems();
-            foreach ($quoteItems as $item) {
-                $item->setData('discount_amount', $item->getOrigData('discount_amount'));
-                $item->setData('base_discount_amount', $item->getOrigData('base_discount_amount'));
-            }
-
-            Mage::helper('boltpay')->collectTotals($quote, true);
-
-            if(!empty($shippingAddressId) && $shippingAddressId != $shippingAddress->getData('address_id')) {
-                $shippingAddress->setData('address_id', $shippingAddressId);
-            }
-        }
-    }
-
-    protected function getSortedShippingRates($address) {
-        $rates = array();
-
-        foreach($address->getGroupedAllShippingRates() as $code => $carrierRates) {
-            foreach ($carrierRates as $carrierRate) {
-                $rates[] = $carrierRate;
-            }
-        }
-
-        return $rates;
-    }
-
-    /**
-     * When Bolt attempts to get shipping rates, it already knows the quote subtotal. However, if there are shipping
-     * methods that could affect the subtotal (e.g. $5 off when you choose Next Day Air), then we need to modify the
-     * shipping amount so that it makes up for the previous subtotal.
-     *
-     * @param float $originalDiscountedSubtotal    Original discounted subtotal
-     * @param Mage_Sales_Model_Quote    $quote    Quote which has been updated to use new shipping rate
-     *
-     * @return float    Discount modified as a result of the new shipping method
-     */
-    public function getAdjustedShippingAmount($originalDiscountedSubtotal, $quote) {
-        return $quote->getShippingAddress()->getShippingAmount() + $quote->getSubtotalWithDiscount() - $originalDiscountedSubtotal;
-    }
-
-    /**
      * Sets Plugin information in the response headers to callers of the API
      */
     public function setResponseContextHeaders()
@@ -1077,7 +933,6 @@ class Bolt_Boltpay_Helper_Api extends Bolt_Boltpay_Helper_Data
             ->setHeader('User-Agent', 'BoltPay/Magento-' . $contextInfo["Magento-Version"], true)
             ->setHeader('X-Bolt-Plugin-Version', $contextInfo["Bolt-Plugin-Version"], true);
     }
-
 
     /**
      * Determines whether the cart has either all items available if Manage Stock is yes for requested quantities,
@@ -1107,7 +962,6 @@ class Bolt_Boltpay_Helper_Api extends Bolt_Boltpay_Helper_Data
         return true;
     }
 
-
     /**
      * Gets a an order by quote id/order reference
      *
@@ -1122,46 +976,5 @@ class Bolt_Boltpay_Helper_Api extends Bolt_Boltpay_Helper_Data
         return $orderCollection
                 ->addFieldToFilter('quote_id', $quoteId)
                 ->getFirstItem();
-    }
-
-    /**
-     * Generate (if) secure url by route and parameters
-     *
-     * @param   string $route
-     * @param   array $params
-     * @return  string
-     */
-    public function getMagentoUrl($route = '', $params = array()){
-        if ((Mage::app()->getStore()->isFrontUrlSecure()) &&
-            (Mage::app()->getRequest()->isSecure())) {
-            $params["_secure"] = true;
-        }
-        return Mage::getUrl($route, $params);
-    }
-
-    /**
-     * Returns user-visible label for given shipping rate.
-     *
-     * @param   object rate
-     * @return  string
-     */
-    public function getShippingLabel($rate) {
-        $carrier = $rate->getCarrierTitle();
-        $title = $rate->getMethodTitle();
-        if (!$title) {
-            return $carrier;
-        }
-
-        // Apply adhoc rules to return concise string.
-        if ($carrier === "Shipping Table Rates") {
-            return $title;
-        }
-        if ($carrier === "United Parcel Service" && substr( $title, 0, 3 ) === "UPS") {
-            return $title;
-        }
-        if (strncasecmp( $carrier, $title, strlen($carrier) ) === 0) {
-            return $title;
-        }
-        return $carrier . " - " . $title;
     }
 }
