+++ app/design/adminhtml/default/default/template/boltpay/system/config/button.phtml
@@ -32,7 +32,7 @@
 
         var xmlhttp = new XMLHttpRequest();
 
-        xmlhttp.open("POST", "<?php echo Mage::helper('boltpay/api')->getMagentoUrl('boltpay/configuration/check');?>", true);
+        xmlhttp.open("POST", "<?php echo Mage::helper('boltpay/url')->getMagentoUrl('boltpay/configuration/check');?>", true);
         xmlhttp.setRequestHeader("Content-type", "application/json");
         xmlhttp.onreadystatechange = function () {
             if (xmlhttp.readyState == XMLHttpRequest.DONE && xmlhttp.status == 200) {
