+++ tests/unit/testsuite/Bolt/Boltpay/Block/Checkout/BoltpayTest.php
@@ -124,7 +124,7 @@ class Bolt_Boltpay_Block_Checkout_BoltpayTest extends PHPUnit_Framework_TestCase
      */
     public function testGetCartURL()
     {
-        $expect = Mage::helper('boltpay/api')->getMagentoUrl('checkout/cart');
+        $expect = Mage::helper('boltpay/url')->getMagentoUrl('checkout/cart');
 
         $result = $this->currentMock->getCartUrl();
 
@@ -149,7 +149,7 @@ class Bolt_Boltpay_Block_Checkout_BoltpayTest extends PHPUnit_Framework_TestCase
     {
         $url = 'checkout/onepage/success';
         $this->app->getStore()->setConfig('payment/boltpay/successpage', $url);
-        $expect = Mage::helper('boltpay/api')->getMagentoUrl($url);
+        $expect = Mage::helper('boltpay/url')->getMagentoUrl($url);
 
         $result = $this->currentMock->getSuccessUrl();
 
@@ -265,7 +265,7 @@ class Bolt_Boltpay_Block_Checkout_BoltpayTest extends PHPUnit_Framework_TestCase
     {
         $this->app->getStore()->setConfig('payment/boltpay/success', '');
         $successCustom = "console.log('test')";
-        $saveOrderUrl = Mage::helper('boltpay/api')->getMagentoUrl('boltpay/order/save');
+        $saveOrderUrl = Mage::helper('boltpay/url')->getMagentoUrl('boltpay/order/save');
         $checkoutType = Bolt_Boltpay_Block_Checkout_Boltpay::CHECKOUT_TYPE_MULTI_PAGE;
 
         $onSuccessCallback = "function(transaction, callback) {
@@ -324,7 +324,7 @@ class Bolt_Boltpay_Block_Checkout_BoltpayTest extends PHPUnit_Framework_TestCase
      */
     public function testBuildOnCloseCallback()
     {
-        $successUrl = Mage::helper('boltpay/api')->getMagentoUrl('checkout/onepage/success');
+        $successUrl = Mage::helper('boltpay/url')->getMagentoUrl('checkout/onepage/success');
         $checkoutType = Bolt_Boltpay_Block_Checkout_Boltpay::CHECKOUT_TYPE_ONE_PAGE;
         $closeCustom = '';
 
@@ -367,7 +367,7 @@ class Bolt_Boltpay_Block_Checkout_BoltpayTest extends PHPUnit_Framework_TestCase
      */
     public function testGetSaveOrderURL()
     {
-        $expect = Mage::helper('boltpay/api')->getMagentoUrl('boltpay/order/save');
+        $expect = Mage::helper('boltpay/url')->getMagentoUrl('boltpay/order/save');
 
         $result = $this->currentMock->getSaveOrderUrl();
