+++ tests/unit/testsuite/Bolt/Boltpay/Helper/ApiTest.php
@@ -121,98 +121,6 @@ class Bolt_Boltpay_Helper_ApiTest extends PHPUnit_Framework_TestCase
         $this->assertEquals($expected, $result);
     }
 
-    /**
-     * Unit test for get adjusted shipping amount with no discount
-     * - subtotal : $100
-     * - shipping amount: $50
-     * - discount amount: none
-     */
-    public function testGetAdjustedShippingAmountWithNoDiscount()
-    {
-        $this->testHelper = new Bolt_Boltpay_TestHelper();
-        $cart = $this->testHelper->addProduct(self::$productId, 2);
-        $quote = $cart->getQuote();
-
-        $originalDiscountedSubtotal = 100;
-
-        $quote->getShippingAddress()->setShippingAmount(50);
-        $quote->setSubtotalWithDiscount(100);
-
-        $expected = 50;
-        $result = $this->currentMock->getAdjustedShippingAmount($originalDiscountedSubtotal, $quote);
-
-        $this->assertEquals($expected, $result);
-    }
-
-    /**
-     * Unit test for get adjusted shipping amount with shipping discount (50%)
-     * - subtotal : $100
-     * - shipping amount: $50
-     * - discount amount: 50% on shipping
-     */
-    public function testGetAdjustedShippingAmountWithShippingDiscount()
-    {
-        $this->testHelper = new Bolt_Boltpay_TestHelper();
-        $cart = $this->testHelper->addProduct(self::$productId, 2);
-        $quote = $cart->getQuote();
-
-        $originalDiscountedSubtotal = 100;
-
-        $quote->getShippingAddress()->setShippingAmount(50);
-        $quote->setSubtotalWithDiscount(75); // Discount on shipping: subtotal - shipping_amount * 50% = $75
-
-        $expected = 25;
-        $result = $this->currentMock->getAdjustedShippingAmount($originalDiscountedSubtotal, $quote);
-
-        $this->assertEquals($expected, $result);
-    }
-
-    /**
-     * Unit test for get adjusted shipping amount with cart discount (50%)
-     * - subtotal : $100
-     * - shipping amount: $50
-     * - discount amount: 50% on quote
-     */
-    public function testGetAdjustedShippingAmountWithCartDiscount()
-    {
-        $this->testHelper = new Bolt_Boltpay_TestHelper();
-        $cart = $this->testHelper->addProduct(self::$productId, 2);
-        $quote = $cart->getQuote();
-
-        $originalDiscountedSubtotal = 50;
-
-        $quote->getShippingAddress()->setShippingAmount(50);
-        $quote->setSubtotalWithDiscount(50); // Discount on quote: subtotal * 50% = $50
-
-        $expected = 50;
-        $result = $this->currentMock->getAdjustedShippingAmount($originalDiscountedSubtotal, $quote);
-
-        $this->assertEquals($expected, $result);
-    }
-
-    /**
-     * Unit test for get adjusted shipping amount with cart discount (50%) and shipping discount (50%)
-     * - subtotal : $100
-     * - shipping amount: $50
-     * - discount amount: 50% on subtotal and 50% on shipping method
-     */
-    public function testGetAdjustedShippingAmountCartAndShippingDiscount()
-    {
-        $this->testHelper = new Bolt_Boltpay_TestHelper();
-        $cart = $this->testHelper->addProduct(self::$productId, 2);
-        $quote = $cart->getQuote();
-
-        $originalDiscountedSubtotal = 50; // Discount on cart: 50%
-
-        $quote->getShippingAddress()->setShippingAmount(50);
-        $quote->setSubtotalWithDiscount(25); // Discount on shipping: subtotal * 50% - shipping_amount * 50% = $25
-
-        $expected = 25;
-        $result = $this->currentMock->getAdjustedShippingAmount($originalDiscountedSubtotal, $quote);
-
-        $this->assertEquals($expected, $result);
-    }
-
     public function testStoreHasAllCartItems()
     {
         $this->testHelper = new Bolt_Boltpay_TestHelper();
@@ -250,57 +158,5 @@ class Bolt_Boltpay_Helper_ApiTest extends PHPUnit_Framework_TestCase
         $result = $this->currentMock->isResponseError($response);
 
         $this->assertTrue($result);
-    }
-
-    public function testShippingLabel()
-    {
-        $rate = $this->getMockBuilder('Mage_Sales_Model_Quote_Address_Rate')
-            ->setMethods(array('getCarrierTitle', 'getMethodTitle'))
-            ->getMock();
-        $rate->method('getCarrierTitle')->willReturn('United Parcel Service');
-        $rate->method('getMethodTitle')->willReturn('2 Day Shipping');
-
-        $label = $this->currentMock->getShippingLabel($rate);
-
-        $this->assertEquals('United Parcel Service - 2 Day Shipping', $label);
-    }
-
-    public function testShippingLabel_notShowShippingTableLatePrefix()
-    {
-        $rate = $this->getMockBuilder('Mage_Sales_Model_Quote_Address_Rate')
-            ->setMethods(array('getCarrierTitle', 'getMethodTitle'))
-            ->getMock();
-        $rate->method('getCarrierTitle')->willReturn('Shipping Table Rates');
-        $rate->method('getMethodTitle')->willReturn('Free shipping (5 - 7 business days)');
-
-        $label = $this->currentMock->getShippingLabel($rate);
-
-        $this->assertEquals('Free shipping (5 - 7 business days)', $label);
-    }
-
-    public function testShippingLabel_notDuplicateCommonPrefix()
-    {
-        $rate = $this->getMockBuilder('Mage_Sales_Model_Quote_Address_Rate')
-            ->setMethods(array('getCarrierTitle', 'getMethodTitle'))
-            ->getMock();
-        $rate->method('getCarrierTitle')->willReturn('USPS');
-        $rate->method('getMethodTitle')->willReturn('USPS Two days');
-
-        $label = $this->currentMock->getShippingLabel($rate);
-
-        $this->assertEquals('USPS Two days', $label);
-    }
-
-    public function testShippingLabel_notDuplicateUPS()
-    {
-        $rate = $this->getMockBuilder('Mage_Sales_Model_Quote_Address_Rate')
-            ->setMethods(array('getCarrierTitle', 'getMethodTitle'))
-            ->getMock();
-        $rate->method('getCarrierTitle')->willReturn('United Parcel Service');
-        $rate->method('getMethodTitle')->willReturn('UPS Business 2 Days');
-
-        $label = $this->currentMock->getShippingLabel($rate);
-
-        $this->assertEquals('UPS Business 2 Days', $label);
     }
 }
